Split solutions into different classes
